New densification methods, variable quad length-scales. increased flexibility for defining quad widths, depths (#81)
* updates to deal with corner cases in acute angle treatment between HUC and river

* generalizing providing river width using any property

* reverting to stream order as standard method, and generalizing using function

- to define the width, any property other than stream order can be defined through function argument, examples, width based on drainage area, labels in node properties, distance from the outlet, etc.

* generalized method to provide river width

* new densify segment that uses LineString interpolation

- this is more robust, especially than the previous method.
- Previous method is available as densify_river old. This will be depreciated eventually once new method is test enough times

* update to end point treatment in redensification

* cleanup in convexity function

* wip commit

* avoid updating of children while looping over

* bug removed

* yapified

* merge functions updated.

For the case when the small segment is one of the tributaries at the junction, we merge it with its child (if there is only one child) and not with its parent.

* cleanup

* typo fixed

* adds option to get multiple subplots in one go from get_ax

* cleanup, NHDPlus source manager temp. patch

* minor updates, trying to get NHD patch work.

* patch mostly working

The property codes in NHD Datasets are in lowercase. The code is updated to change property codes to lowercase in cases the new format of the NHD Dataset is encountered. This change is applicable for not only 'NHDFlowline' layer, but also in layers 'NHDPlusCatchment', 'NHDPlusFlowlineVAA',
 NHDPlusEROMMA'.

* bug removed, the NHDPlus data/property working now

* tests updated

Two tests were failing; this could be due to updates in the hydrography in the new NHD dataset. All other tests are passing fine, and all example problems are also running fine.

FAILED watershed_workflow/test/test_hilev.py::test_river_tree_properties - assert 94 == 97
FAILED watershed_workflow/test/test_hilev.py::test_river_tree_properties_prune - assert 49 == 50

* update to merging: option 1 for merging into child

* update to merging: option 2 for merging into child

We just keep the node to be merged as self, and , merge it into parent or child based on a flag.

* updates unit tests

* yapf

---------

Co-authored-by: Ethan Coon <coonet@ornl.gov>