cleaning up densification, and major name changes (see below)
Some standardization was needed for names across the project.  The
following variable names are now used:

river = RiverTree object when the tree is needed
reach = RiverTree object when the tree is NOT needed
reaches = GeoDataFrame of linestrings + properties is still ok
linestring = RiverTree.linestring is the geometry of the reach.  Note
    this used to be called "segment" which was a bad name, as segments
    are subsets of  _line_ with two endpoints
segment = is now used for the subset of a line defined by two points
    Note this used to be used this way in utils.

Also, densification was rename resampling, since it no longer just
densifies points.

So, as an example, one could:

- loops over a list of rivers (trunk nodes)
   - for each reach in the river (preOrder)
       - resample(reach.linestring)
          - for each segment in reach:
              - densify(segment)

Hopefully using this standard convention everywhere in the code will
help clarify.