New logic for env and main CI tasks and docker images
1. runs env.yml task on PR to master and geopandas_xarray
2. main.yml searches for a ci-env image on dockerhub,
or defaults to master tag if it is not there.

Note because of the multiarch changes, the env.yml
task will have to run once on each branch
before the main task. Might be worth running these
in a single CI task, but where the env steps only
run if there are changes to certain files in the
repository (similar to how Amanzi-ATS TPLs are handled
now in amanzi/amanzi)