Update CI for new environments, packages (#121)
* redirect stderr when checking for ci-env image

* update create_envs.py calls based on current args

* add seacas patch for newer netcdf

* adjust tools env for multiarch

* not the most elegant patches to build pip packages and seacas but works

* get correct name of secret in yaml

* harmonize tags across ci jobs

* Specify cross-architecture in fix-manifest step

Not sure if this is due to buildx or a change
in the runner docker version, but the fix manifest
step is not working because it will not checkout
the image on a different architecture by default.
Trying to manually specify the platform arg for
cross-architecture to make sure docker pull grabs
both imges

* Try emulated approach

buildx even when building for one architecture
apparently creates a manifest list, which causes
issues when trying to stitch together a multiarch
image.

i had been trying to avoid the emulator since it is slow,
but fairly little code gets compiled here and the
job is pretty short overall, so trying emulation
through QEMU

* fix branch tag test logic that always evaluates true

* fix branch tag in build arg

* Add LD_LIBRARY_PATH to CI-Env container

CI container fails because libsqlite3.so cannot
be found (and, it still carries the old path of
/opt/conda/envs/watershed_workflow_CI/lib). Trying
to set LD_LIBRARY_PATH to change the search order,
but not sure how it is searching /opt/conda unless
there are some library linkages that are not
updated by conda_pack/conda_unpack

* rework CI conda environments

* remove latest from tags

* fix rioxarray deprecation warnings

* fix missed conflict in manager_modis_appeears