Major refactor of sources concept (#128)
This refactors both Dataset and Shape managers, creating a base class for each, ensuring a common API across all such classes. It also updates SoilGrids to modern manager design. Manager tests should now also work with CI, and are protected from submitting large downloads.

This refactor also opens the opportunity for delayed, non-blocking requests for data. See ManagerDataset for the fundamentals, where a Request (conceptually a Future) is created and returned after requestDataset(), and then finalized by a wait() call or by isReady() followed by fetchDataset(). Note that few managers use the nonblocking strategy, but the goal is to eventually refactor them to launch subprocesses to download the data and return immediately, then allow fetching the data when the download isReady().

It also updates a bunch of notebooks to work with the new version.