Workaround attempt for pymetis/metis arm64 bug
Somewhere in the feedstock for pymetis on conda-forge, something is
messed up when doing an arm64 build.  pymetis's _internal cython
library ends up with the amd64 internal metis library.  I'm not sure
where this goes wrong.

pip install pymetis (in an arm64 arch) gets the right _internal
library.

I'm not sure if this is a metis problem, a pymetis problem, a
feedstock problem, or what.  For now, let's try explicitly getting
metis -- if this doesn't work, will try adding pymetis to the
requirements.txt instead.